src/main.py
import json
import os
from dotenv import load_dotenv

load_dotenv()

import datetime
from scrapers import get_rss_news, get_hacker_news, get_reddit_news, get_arxiv_papers
from summarizer import summarize_content

def main():
    print("Starting AI Daily News aggregation...")
    
    all_news = []
    
    # 1. RSS Feeds
    rss_urls = [
        'https://openai.com/blog/rss.xml',
        'https://deepmind.google/blog/rss.xml',
        'https://huggingface.co/blog/feed.xml'
    ]
    print("Fetching RSS feeds...")
    all_news.extend(get_rss_news(rss_urls))
    
    # 2. Hacker News
    print("Fetching Hacker News...")
    all_news.extend(get_hacker_news())
    
    # 3. Reddit
    print("Fetching Reddit...")
    all_news.extend(get_reddit_news())
    
    # 4. arXiv
    print("Fetching arXiv...")
    all_news.extend(get_arxiv_papers())
    
    print(f"Collected {len(all_news)} items. Deduplicating...")
    
    # Deduplicate by URL
    seen_urls = set()
    unique_news = []
    for item in all_news:
        if item['url'] not in seen_urls:
            seen_urls.add(item['url'])
            unique_news.append(item)
            
    print(f"Unique items: {len(unique_news)}. Summarizing (this may take a while)...")
    
    # Summarize top 10 items to save API costs/time for demo, or all if production
    # For this implementation, let's limit to top 15 most interesting looking ones or just first 15
    # In a real app, we might score them. Here we just take the first 15 mixed from sources.
    
    final_news = []
    for i, item in enumerate(unique_news[:50]):
        print(f"Summarizing {i+1}/15: {item['title']}")
        # We pass the title and maybe a snippet if we had it. 
        # Since we only have title/url mostly, we'll ask the LLM to summarize based on title/context 
        # OR we would need to scrape the content. 
        # For MVP, we'll ask LLM to "Explain why this might be important based on the title" 
        # or if we can, we fetch the page content. 
        # Fetching page content is heavy. Let's stick to title-based "Why it matters" or just clean up the title.
        # WAIT: The user wants a "Summarization script". 
        # Ideally we fetch the content. Let's try to fetch text content for better summaries.
        
        try:
            # Simple content fetch for context (timeout short)
            # This is risky without a headless browser for some sites, but okay for MVP
            # If it fails, we fall back to title.
            content_context = f"Title: {item['title']}\nSource: {item['source']}"
            
            summary = summarize_content(content_context)
            item['summary'] = summary
            final_news.append(item)
        except Exception as e:
            print(f"Failed to summarize {item['title']}: {e}")
            item['summary'] = "Summary unavailable."
            final_news.append(item)
            
    # Save to data/news.json
    output_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'data', 'news.json')
    with open(output_path, 'w') as f:
        json.dump({
            'updated': datetime.datetime.now().isoformat(),
            'articles': final_news
        }, f, indent=2)
        
    print(f"Done! Saved {len(final_news)} articles to {output_path}")

if __name__ == "__main__":
    main()
