requirements.txt
requests
beautifulsoup4
feedparser
groq
openai
python-dotenv
